Runner at   runners/generation/run_generation.py, entrypoint-style.
  - fews_agent/generators/base.py: render(), canonicalize() (uses
    remove_blank_text + remove_comments on the parser), GeneratorSpec
    dataclass, and an xmlstr Jinja filter that forces Decimal to fixed-point
    notation so 0.00000000000000000001 doesn't collapse to 1E-20.
  - fews_agent/generators/{locations,parameters,qualifiers}.py: one thin
    generate() per type; SPECS registry in __init__ is the authoritative
    list of producible file types.
  - fews_agent/generators/templates/*.xml.j2: one template per type.
  - runners/generation/run_generation.py: iterates SPECS, writes to
    examples/generated-config-tutorial/, C14N-compares vs tutorial, prints
    [OK]/[FAIL] per file with a unified diff on mismatch. json.load uses
    parse_float=Decimal so source digits survive.

+++ fews_agent/schema/locations.py
@@ -1,6 +1,8 @@
 """Locations.xml — declares locationId used across the config."""
 from __future__ import annotations
 
+from decimal import Decimal
+
 from pydantic import Field
 
 from .common import FewsModel
@@ -13,13 +15,17 @@ class LocationAttribute(FewsModel):
 
 
 class Location(FewsModel):
+    # Decimal (not float) preserves the exact numeric representation from
+    # the JSON source. Integer JSON values emit as "-180" (not "-180.0"),
+    # and long decimals like 0.00000000000000000001 survive intact. Required
+    # for C14N equality against hand-authored XML.
     id: LocationId
     name: str
-    x: float
-    y: float
+    x: Decimal
+    y: Decimal
     shortName: str | None = None
     description: str | None = None
-    z: float | None = None
+    z: Decimal | None = None
     parentLocationId: LocationId | None = None
     relation: str | None = None
     attribute: list[LocationAttribute] = Field(default_factory=list)
