+++ fews_agent/schema/parameters.py
@@ -1,6 +1,8 @@
 """Parameters.xml — declares parameterId and parameterGroupId."""
 from __future__ import annotations
 
+from decimal import Decimal
+
 from pydantic import Field
 
 from .common import FewsModel
@@ -17,11 +19,14 @@ class Parameter(FewsModel):
 
 
 class ParameterGroup(FewsModel):
+    # valueResolution uses Decimal to preserve the exact source text from
+    # JSON (e.g. "0.00000000000000000001" must not collapse to "1E-20").
+    # See comment in schema.locations re: Decimal.
     id: ParameterGroupId
     parameterType: ParameterType
     unit: str
     name: str | None = None
-    valueResolution: float | None = None
+    valueResolution: Decimal | None = None
     usesDatum: bool = False
     parameter: list[Parameter] = Field(min_length=1)
